Boardクラスのファイル切り出し
